pyproject.toml
[tool.poetry]
name = "llm-promt-injection-detector"
version = "0.1.0"
description = ""
authors = ["Dimitry Perchatkin <perchatkindimitry@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = "^0.30.1"
pydantic = "^2.8.0"
google-generativeai = "^0.7.1"
configparser = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
